refactor: split HistoryEventsTable.vue
